Add language codes for the final list of sites.
We ran a crawl to detect the language for 46568 shopping websites.

Due to timeouts, selenium errors and encoding related issues we could not
detect the language on 4190 sites. We should include those sites
we fail to detect in the final crawl.

Data for the language detection crawl is at:
`/n/fs/darkpatterns/crawl/crawl-all-shopping-lang-detection-close-dialog-181005`