Only show subset of each cluster in cluster browser for speed
It takes a long time to read in the text of large clusters (e.g.
with thousands of segments), and it's not really necessary to
read more than a few 10s of segments to get a sense of whether
the cluster is useful or not anyway. So we only show a small
subset of segments from each cluster.